refactor(btrblocks): split scheme files into per-scheme modules
The integer, float, and string scheme files each contained many distinct
Scheme implementations in a single module. Promote each to a directory
module with one submodule per scheme, re-exporting the scheme types from
the parent module so the public paths (e.g. schemes::integer::FoRScheme)
are unchanged.

- integer/: bitpacking, frame_of_reference, rle, runend, sequence, sparse,
  zigzag, pco
- float/: alp, alprd, rle, sparse, pco
- string/: fsst, sparse, zstd, zstd_buffers, onpair

Shared helpers (rle_compress, try_compress_delta) and RUN_LENGTH_THRESHOLD
move into the integer rle submodule and are re-exported at the integer
module root. Tests move alongside their module into tests.rs and
scheme_selection_tests.rs files.
