+++ vortex-btrblocks/src/schemes/float/alprd.rs
@@ -1,0 +1,89 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+//! ALPRD (ALP with Real Double) encoding variant.
+
+use vortex_alp::ALPRDArrayExt;
+use vortex_alp::ALPRDArrayOwnedExt;
+use vortex_alp::RDEncoder;
+use vortex_array::ArrayRef;
+use vortex_array::Canonical;
+use vortex_array::ExecutionCtx;
+use vortex_array::IntoArray;
+use vortex_array::arrays::primitive::PrimitiveArrayExt;
+use vortex_array::dtype::PType;
+use vortex_compressor::estimate::CompressionEstimate;
+use vortex_compressor::estimate::DeferredEstimate;
+use vortex_compressor::estimate::EstimateVerdict;
+use vortex_error::VortexResult;
+use vortex_error::vortex_panic;
+
+use crate::ArrayAndStats;
+use crate::CascadingCompressor;
+use crate::CompressorContext;
+use crate::Scheme;
+use crate::compress_patches;
+
+/// ALPRD (ALP with Real Double) encoding variant.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct ALPRDScheme;
+
+impl Scheme for ALPRDScheme {
+    fn scheme_name(&self) -> &'static str {
+        "vortex.float.alprd"
+    }
+
+    fn matches(&self, canonical: &Canonical) -> bool {
+        canonical.dtype().is_float()
+    }
+
+    fn expected_compression_ratio(
+        &self,
+        data: &ArrayAndStats,
+        _compress_ctx: CompressorContext,
+        _exec_ctx: &mut ExecutionCtx,
+    ) -> CompressionEstimate {
+        // We don't support ALPRD for f16.
+        if data.array_as_primitive().ptype() == PType::F16 {
+            return CompressionEstimate::Verdict(EstimateVerdict::Skip);
+        }
+
+        CompressionEstimate::Deferred(DeferredEstimate::Sample)
+    }
+
+    fn compress(
+        &self,
+        _compressor: &CascadingCompressor,
+        data: &ArrayAndStats,
+        _compress_ctx: CompressorContext,
+        exec_ctx: &mut ExecutionCtx,
+    ) -> VortexResult<ArrayRef> {
+        let primitive_array = data.array_as_primitive();
+
+        let encoder = match primitive_array.ptype() {
+            PType::F32 => RDEncoder::new(primitive_array.as_slice::<f32>()),
+            PType::F64 => RDEncoder::new(primitive_array.as_slice::<f64>()),
+            ptype => vortex_panic!("cannot ALPRD compress ptype {ptype}"),
+        };
+
+        let alp_rd = encoder.encode(primitive_array, exec_ctx);
+        let dtype = alp_rd.dtype().clone();
+        let right_bit_width = alp_rd.right_bit_width();
+        let mut parts = ALPRDArrayOwnedExt::into_data_parts(alp_rd);
+        parts.left_parts_patches = parts
+            .left_parts_patches
+            .map(|p| compress_patches(p, exec_ctx))
+            .transpose()?;
+
+        Ok(vortex_alp::ALPRD::try_new(
+            dtype,
+            parts.left_parts,
+            parts.left_parts_dictionary,
+            parts.right_parts,
+            right_bit_width,
+            parts.left_parts_patches,
+            exec_ctx,
+        )?
+        .into_array())
+    }
+}
